Split store from components
